rename walter_command.py to command_line.py
